[SECURITY] Disable HMAC sig methods by default due to key confusion
Both HMAC and public key digital signature methods use the same
variable to provide the key, and the choice of algorithm is provided
by user input. This can result in a security vulnerability. Since
HMAC-SHA1 is less commonly use, disable it by default to prevent
this issue. If you use it, you can re-enable it by calling
SignedXml.enableHMAC().

Since this is a breaking change to the API, update major version
number to 2.0.0